[DEVOPS-1199] Don't use a separate cabal package for cardano-wallet-client
