--- wallet-tool/Setup.hs
@@ -1,4 +1,0 @@
-#! /usr/bin/env nix-shell
-#! nix-shell ./default.nix -i runghc
-import           Distribution.Simple
-main = defaultMain
