--- wallet-tool/cardano-wallet-client.cabal
@@ -1,174 +1,0 @@
-name:                cardano-wallet-client
-version:             2.0.0
-synopsis:            A CLI to making API calls on the wallet
-description:         Please see README.md
-homepage:            https://github.com/input-output-hk/cardano-wallet
-author:              IOHK Engineering Team
-maintainer:          operations@iohk.io
-copyright:           2019 IOHK
-license:             MIT
-license-file:        LICENSE
-category:            Web
-build-type:          Simple
-extra-source-files:  README.md
-cabal-version:       >=1.10
-
-library
-  hs-source-dirs:      src
-  exposed-modules:     Cardano.Wallet.Client.Easy
-                     , Cardano.Wallet.Client.Wait
-                     , Cardano.Wallet.ProcessUtil
-                     , Cardano.Wallet.Client.HttpsSettings
-                     , Cardano.Wallet.Client.CLI
-  other-modules:       Paths_cardano_wallet_tool
-  ghc-options:         -Wall
-
-  build-depends:       base
-                     , aeson
-                     , aeson-pretty
-                     , async
-                     , attoparsec
-                     , bytestring
-                     , cardano-sl
-                     , cardano-sl-client
-                     , cardano-sl-core
-                     , cardano-sl-mnemonic
-                     , cardano-sl-x509
-                     , cardano-wallet
-                     , connection
-                     , criterion
-                     , data-default
-                     , dlist
-                     , formatting
-                     , filepath
-                     , http-client
-                     , http-client-tls
-                     , optparse-applicative
-                     , parsec
-                     , retry
-                     , safe-exceptions
-                     , servant-client
-                     , servant-client-core
-                     , stm
-                     , text
-                     , time
-                     , tls
-                     , universum
-                     , x509
-                     , x509-store
-                     , pem
-
-  if !os(windows)
-    build-depends:     unix
-
-  default-language:    Haskell2010
-  default-extensions: TypeOperators
-                      DataKinds
-                      DefaultSignatures
-                      NoImplicitPrelude
-                      MultiParamTypeClasses
-                      OverloadedStrings
-                      ScopedTypeVariables
-                      FlexibleInstances
-                      FlexibleContexts
-                      TypeFamilies
-                      TypeApplications
-                      TypeOperators
-                      TemplateHaskell
-                      RecordWildCards
-                      ScopedTypeVariables
-                      UndecidableInstances
-                      MonadFailDesugaring
-                      TupleSections
-
-executable cardano-wallet-client
-  hs-source-dirs:      exe
-  main-is:             Main.hs
-  ghc-options:         -threaded -rtsopts
-                       -Wall
-                       -O2
-
-  build-depends:       base
-                     , cardano-wallet-client
-                     , optparse-applicative
-                     , universum
-
-  default-language:    Haskell2010
-  default-extensions: TypeOperators
-                      DataKinds
-                      DefaultSignatures
-                      NoImplicitPrelude
-                      MultiParamTypeClasses
-                      OverloadedStrings
-                      ScopedTypeVariables
-                      FlexibleInstances
-                      FlexibleContexts
-                      TypeFamilies
-                      TypeApplications
-                      TypeOperators
-                      TemplateHaskell
-                      RecordWildCards
-                      ScopedTypeVariables
-                      UndecidableInstances
-                      MonadFailDesugaring
-                      TupleSections
-
-executable cardano-wallet-sync-plot
-  hs-source-dirs:      exe
-  main-is:             sync-plot.hs
-  ghc-options:         -threaded -rtsopts
-                       -Wall
-                       -O2
-
-  build-depends:       base
-                     , optparse-applicative
-                     , aeson
-                     , Chart
-                     , Chart-cairo
-                     , text
-                     , universum
-                     , vector
-
-  default-language:    Haskell2010
-  default-extensions: NoImplicitPrelude
-                      OverloadedStrings
-                      RecordWildCards
-
-test-suite wallet-tool-specs
-  ghc-options:      -Wall -O2 -threaded -rtsopts
-  type:             exitcode-stdio-1.0
-  main-is:          Spec.hs
-  hs-source-dirs:   test
-  other-modules:    WaitSpec
-
-  default-language: Haskell2010
-  default-extensions: BangPatterns
-                      ConstraintKinds
-                      DeriveGeneric
-                      FlexibleContexts
-                      FlexibleInstances
-                      GADTs
-                      InstanceSigs
-                      LambdaCase
-                      MultiParamTypeClasses
-                      MultiWayIf
-                      NoImplicitPrelude
-                      OverloadedStrings
-                      RankNTypes
-                      RecordWildCards
-                      ScopedTypeVariables
-                      StandaloneDeriving
-                      TypeFamilies
-                      TypeApplications
-                      TypeOperators
-                      DataKinds
-                      TemplateHaskell
-
-  build-depends:      base
-                    , cardano-sl
-                    , cardano-wallet-client
-                    , cardano-wallet
-                    , hspec
-                    , servant-client
-                    , text
-                    , universum
