--- wallet-tool/test/Spec.hs
@@ -1,11 +1,0 @@
-module Main where
-
-import           Universum
-
-import           Test.Hspec
-
-import qualified WaitSpec
-
-main :: IO ()
-main = do
-    hspec $ WaitSpec.spec
